breaking change: audio sources are always floating point, removed audiorenderer audio decoding responsibility, fixed audio duration calculations, added audio device switching

+++ Walgelijk.OpenTK/Audio/AudioFileData.cs
@@ -4,6 +4,11 @@ public struct AudioFileData
 {
     public short NumChannels;
     public int SampleRate;
+
+    /// <summary>
+    /// Sample count per chanel
+    /// </summary>
     public long SampleCount;
-    public byte[]? Data;
+
+    public float[]? Data;
 }
